+++ src/parser/giganto_import/network.rs
@@ -7,8 +7,9 @@ use giganto_client::ingest::network::{
 };
 
 use super::{
-    GigantoImportResult, TryFromGigantoRecord, parse_comma_separated, parse_giganto_timestamp_ns,
-    parse_parenthesized_tuples, parse_post_body,
+    GigantoImportResult, TryFromGigantoRecord, parse_comma_separated,
+    parse_embedded_giganto_datetime_ns, parse_giganto_timestamp_ns, parse_parenthesized_tuples,
+    parse_post_body,
 };
 
 type Result<T> = GigantoImportResult<T>;
@@ -64,7 +65,7 @@ impl TryFromGigantoRecord for Conn {
             return Err(giganto_import_error!("missing conn state"));
         };
         let start_time = if let Some(start_time) = rec.get(8) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -186,7 +187,7 @@ impl TryFromGigantoRecord for Dns {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -380,7 +381,8 @@ impl TryFromGigantoRecord for MalformedDns {
             .context("missing protocol")?
             .parse::<u8>()
             .context("invalid proto")?;
-        let start_time = parse_giganto_timestamp_ns(rec.get(7).context("missing start_time")?)?;
+        let start_time =
+            parse_embedded_giganto_datetime_ns(rec.get(7).context("missing start_time")?)?;
         let duration = rec
             .get(8)
             .context("missing duration")?
@@ -530,7 +532,7 @@ impl TryFromGigantoRecord for Http {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -751,7 +753,7 @@ impl TryFromGigantoRecord for Rdp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -853,7 +855,7 @@ impl TryFromGigantoRecord for Smtp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -990,7 +992,7 @@ impl TryFromGigantoRecord for Ntlm {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1116,7 +1118,7 @@ impl TryFromGigantoRecord for Kerberos {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1152,12 +1154,12 @@ impl TryFromGigantoRecord for Kerberos {
             return Err(giganto_import_error!("missing destination l2 bytes"));
         };
         let client_time: i64 = if let Some(client_time) = rec.get(13) {
-            parse_giganto_timestamp_ns(client_time)?
+            parse_embedded_giganto_datetime_ns(client_time)?
         } else {
             return Err(giganto_import_error!("missing client_time"));
         };
         let server_time: i64 = if let Some(server_time) = rec.get(14) {
-            parse_giganto_timestamp_ns(server_time)?
+            parse_embedded_giganto_datetime_ns(server_time)?
         } else {
             return Err(giganto_import_error!("missing server_time"));
         };
@@ -1271,7 +1273,7 @@ impl TryFromGigantoRecord for Ssh {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1444,7 +1446,7 @@ impl TryFromGigantoRecord for DceRpc {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1570,7 +1572,7 @@ impl TryFromGigantoRecord for Ftp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1752,7 +1754,7 @@ impl TryFromGigantoRecord for Mqtt {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -1885,7 +1887,7 @@ impl TryFromGigantoRecord for Ldap {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -2037,7 +2039,7 @@ impl TryFromGigantoRecord for Tls {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -2136,15 +2138,13 @@ impl TryFromGigantoRecord for Tls {
             return Err(giganto_import_error!("missing subject_common_name"));
         };
         let validity_not_before = if let Some(validity_not_before) = rec.get(26) {
-            validity_not_before
-                .parse::<i64>()
+            parse_embedded_giganto_datetime_ns(validity_not_before)
                 .context("invalid validity_not_before")?
         } else {
             return Err(giganto_import_error!("missing validity_not_before"));
         };
         let validity_not_after = if let Some(validity_not_after) = rec.get(27) {
-            validity_not_after
-                .parse::<i64>()
+            parse_embedded_giganto_datetime_ns(validity_not_after)
                 .context("invalid validity_not_after")?
         } else {
             return Err(giganto_import_error!("missing validity_not_after"));
@@ -2260,7 +2260,7 @@ impl TryFromGigantoRecord for Smb {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -2422,7 +2422,7 @@ impl TryFromGigantoRecord for Nfs {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -2536,7 +2536,7 @@ impl TryFromGigantoRecord for Bootp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -2696,7 +2696,7 @@ impl TryFromGigantoRecord for Dhcp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -3004,7 +3004,7 @@ impl TryFromGigantoRecord for Radius {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(7) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };
@@ -3158,7 +3158,7 @@ impl TryFromGigantoRecord for Icmp {
             return Err(giganto_import_error!("missing protocol"));
         };
         let start_time = if let Some(start_time) = rec.get(5) {
-            parse_giganto_timestamp_ns(start_time)?
+            parse_embedded_giganto_datetime_ns(start_time)?
         } else {
             return Err(giganto_import_error!("missing start_time"));
         };